move types to new types.go file
